requirements.txt
# dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
PyJWT>=2.8.0
cryptography>=42.0.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
